Strengthen claim that b*exp(a*t) is general soln
Previously it is claimed that for any value of b the classic dy/dt=a*y
is the above soln. This is true, but why not mention that this ODE is
satisfied for any value of a, too? This re-inforces (but does not prove)
the claim that the above is the general solution to the ODE.