Update to-do list
Removed the item about adding figures for dot and cross product: the
discussion in the book is so cursory, such that it would only really be
valuable for people already familiar with these operations.

Removed the item about an answer key. The code included with the book is
alreaedy an answer key (albeit incomplete).